requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openpyxl>=3.1.0
python-docx>=0.8.11
requests>=2.31.0
